requirements.txt
annotated-types==0.7.0
APScheduler==3.11.2
asgiref==3.11.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
Django==6.0.4
django-apscheduler==0.7.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
dotenv==0.9.9
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
httplib2==0.31.2
idna==3.11
packaging==26.1
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
soupsieve==2.8.3
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
